[FIX] Moved store to _core folder
