+++ src/app/_core/store/cart/cart.actions.ts
@@ -1,6 +1,5 @@
 import { createAction, props } from '@ngrx/store';
-import { CartItem } from './cart';
-import { Product } from '../../_core/interfaces/product';
+import { Product } from '../../interfaces/product';
 
 export const loadCart = createAction('[Cart] Load Cart');
 export const addToCart = createAction('[Cart] Add To Cart', props<{ item: Product }>());
