+++ src/app/app.config.ts
@@ -1,12 +1,11 @@
 import { ApplicationConfig } from '@angular/core';
 import { provideRouter } from '@angular/router';
-
 import { routes } from './app.routes';
 import { HTTP_INTERCEPTORS, provideHttpClient, withInterceptors, withInterceptorsFromDi } from '@angular/common/http';
 import { provideStore } from '@ngrx/store';
-import { cartReducer } from './store/cart/cart.reducers';
 import { ResponseInterceptor } from './_core/interceptors/response.interceptor';
 import { RequestInterceptor } from './_core/interceptors/request.interceptor';
+import { cartReducer } from './_core/store/cart/cart.reducers';
 
 export const appConfig: ApplicationConfig = {
   providers: [
